server/pyproject.toml
[tool.poetry]
name = "mindreview"
version = "0.1.0"
description = ""
authors = ["AlekseyAmp <thientaovv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
flake8 = "^7.0.0"
autopep8 = "^2.0.4"
pytz = "^2024.1"
fastapi-jwt-auth = {extras = ["asymmetric"], version = "^0.5.0"}
passlib = "^1.7.4"
fastapi = "0.99.1"

[tool.isort]
profile = "black"
known_first_party = "kgok_api"
known_evraz = "evraz"
sections=["FUTURE", "STDLIB", "THIRDPARTY", "EVRAZ", "FIRSTPARTY", "LOCALFOLDER"]
line_length = 80

[tool.flake8]
max-line-length = 80
select = ['F', 'E', 'W', ]

ignore = ['W503', ]

per-file-ignores = [
    '*/adapters/*/repositories.py:E711,E712',
    '*/adapters/*/repositories/*:E711,E712',
    '*/migrations/*:E711,E712',
    '*/alembic/versions/*:E711,E712',
    '__init__.py:F401',
]
count = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

